Add a symbol to represent for items in a shop in console mode.
If you start Crawl with tiles, there were already images of the stock in a shop. This change means that console mode has glyphs along the same lines.

Hide the random appearance of unidentified items which are in a shop which identifies things, such as a general store. An unidentified wand will have a wand glyph next to it, but the glyph will not usually match the wand in the description.

With default settings, the game also hides the appearance of unidentified items in antique shops. This does not leak information, and the player can change the behaviour.