Only list hints in "examine surroundings" for keys which do something.
describe (v) is shown if there is a known (or remembered) thing on the square.

travel (.) is shown if the wall is known (or remembered) as a non-wall square. Entering the select key no longer leaves do_look_around() where this isn't the case.

get item (g) is shown if a known item is shown on a square with known terrain.

Previously, v was only shown for squares which are in the player's line of sight, . was shown for all locations, and g for items and detected items.