Warn the user if a TileColGameOption is set to a bad value.
Previously, the game silently set the relevant option to another colour (usually black) if it could not parse the requested value. It now reports an error and leaves the option unchanged.

Use strtol to decode hexadecimal codes for this (instead of a hand-rolled function) and separate out the list of colours from the code details.