Create a "string" GameOption with a custom "set this option" functions.
Use it for the "fire_items_start" option. This now rejects too-long strings rather than ignoring the ends of them.

Nothing else should change the underlying variables for an option because CustomStringGameOption::str() doesn't read them. Store the variable as a const reference to a private member variable to enable the compiler to check this.