Create a GameOption for lists of data which use a special parser.
This class maintains a separate structure to represent the variable instead of examining the game state, so the option should only be set via this class.

"-=" changes are implemented by adding a line with a space at the start of it. The actual effect may vary from one option to another.

Change display_char to use this class. To match the existing behaviour, every setting for this is treated as being set with "+=" as the line type.