Change species, background, weapon and combo into CustomListGameOptions.
Choose allowed species, jobs & weapons from the start-ns.prf file at random, so that the order of settings is not significant. The game doesn't add more than one setting to this file for each option, but the user may do.

"weapon ^= xxx" now adds to the start of the list rather than (as before) the end. This does not affect what is chosen.

Add a use_combo option to show that species/background/weapon combinations should be used instead of relying on allowed_combos being empty. This is set when one of the lists changes.

Add a "if option A has changed option B may also have changed" list. Use this to ensure that use_combo is displayed correctly as the options are edited.