Add help feature to the options menu and some submenus.
If the user presses ? in an options menu, bring up the part of options_guide.txt, arena.txt or fight_simulator.txt which describes that option. The latter two are technically now available in the help menu, but the player isn't told where they are.

Also allow it within "choose from a list" and "edit a list of parameters" options. The "edit a string" options don't do this, as a ? is just a ? there.