Arrange the options according to the help files.
Add a function to list the options in option_behaviour in order of where each appears in options_guide.txt (or other files), and add subheadings for the sections. List options which aren't listed in the files separately.

Put the "show help" function and data in GameOption.

Fix an option name in options_guide.txt (the index was already correct).