Improve the option setting experience for ListGameOptions.
Give ListGameOption::load_from_UI() an editor which lets the user add,
delete, alter and rearrange the values in the list.

The new function uses vector<string> input and output, so make string
and text_pattern child classes to convert between them.

Delete the << override for text_pattern, which is no longer needed.